+++ changelog.md
@@ -4,6 +4,12 @@ _(Note: This changelog focuses on the major changes between the different
 versions. Therefore, it may not contain all changes. Especially smaller fixes or
 improvements may be omitted.)_
 
+## Version 0.?.? (2024-03-??)
+
+* __[fix]__
+  The `!fortune` command will now work on Linux distributions that are not based
+  on Debian, too.
+
 ## Version 0.8.0 (2024-03-01)
 
 * __[change]__